requirements.txt
Django==2.2.4
django-ckeditor==5.8.0
django-js-asset==1.2.2
django-tinymce==2.8.0
Pillow==7.0.0
pytz==2019.3
sqlparse==0.3.0